pipelines/load-featurestore-pipeline/create-entity/pyproject.toml
[tool.poetry]
name = "load"
version = "0.1.0"
description = ""
authors = ["tonouchi27 <tonouchi27@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9,<3.10"
absl-py = "^1.0.0"
PyYAML = "^6.0"
google-cloud-aiplatform = "^1.8.0"
google-cloud-bigquery = "^2.31.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
